+++ src/fields/OptimizedImages.php
@@ -202,6 +202,10 @@ class OptimizedImages extends Field
                         $transform,
                         ImageOptimize::$transformParams
                     );
+                    Craft::info(
+                        'URL created: '.print_r($url, true),
+                        __METHOD__
+                    );
                     // Update the model
                     if (!empty($url)) {
                         // Store & prefetch image at the image URL
@@ -223,11 +227,19 @@ class OptimizedImages extends Field
                         $this->generatePlaceholders($element, $model, $aspectRatio);
                         $placeholderMade = true;
                     }
+                    Craft::info(
+                        'Created transforms for variant: '.print_r($variant, true),
+                        __METHOD__
+                    );
+                } else {
+                    Craft::error(
+                        'Could not create transform for: '.$element->title
+                            . " - Final format: ".$finalFormat
+                            . " - Element extension: ".$element->getExtension()
+                            . " - canManipulateAsImage: ".Image::canManipulateAsImage($element->getExtension()),
+                        __METHOD__
+                    );
                 }
-                Craft::info(
-                    'Created transforms for variant: '.print_r($variant, true),
-                    __METHOD__
-                );
             }
         }
     }