+++ src/services/Optimize.php
@@ -16,6 +16,7 @@ use nystudio107\imageoptimize\fields\OptimizedImages;
 use Craft;
 use craft\base\Component;
 use craft\base\Image;
+use craft\base\LocalVolumeInterface;
 use craft\base\Volume;
 use craft\elements\Asset;
 use craft\errors\VolumeException;
@@ -173,7 +174,7 @@ class Optimize extends Component
     protected function executeImageProcessor($thisProcessor, string $tempPath)
     {
         // Make sure the command exists
-        if (file_exists($thisProcessor['commandPath'])) {
+        if (is_file($thisProcessor['commandPath'])) {
             // Set any options for the command
             $commandOptions = '';
             if (!empty($thisProcessor['commandOptions'])) {
@@ -335,7 +336,7 @@ class Optimize extends Component
     {
         $outputPath = $tempPath;
         // Make sure the command exists
-        if (file_exists($variantCreatorCommand['commandPath'])) {
+        if (is_file($variantCreatorCommand['commandPath'])) {
             // Get the output file for this image variant
             $outputPath .= '.'.$variantCreatorCommand['imageVariantExtension'];
             // Set any options for the command
@@ -398,7 +399,7 @@ class Optimize extends Component
         $outputPath
     ) {
         // If the image variant creation succeeded, copy it into place
-        if (!empty($outputPath) && file_exists($outputPath)) {
+        if (!empty($outputPath) && is_file($outputPath)) {
             // Figure out the resulting path for the image variant
             $volume = $asset->getVolume();
             $assetTransforms = Craft::$app->getAssetTransforms();
@@ -412,17 +413,28 @@ class Optimize extends Component
                 // We're fine with that.
             }
 
+            Craft::info(
+                'Variant output path: '.$outputPath.' - Variant path: '.$variantPath,
+                __METHOD__
+            );
+
             clearstatcache(true, $outputPath);
-            $stream = fopen($outputPath, 'rb');
+            $stream = @fopen($outputPath, 'rb');
 
             // Now create it
             try {
                 $volume->createFileByStream($variantPath, $stream, []);
             } catch (VolumeObjectExistsException $e) {
                 // We're fine with that.
+                Craft::error(
+                    Craft::t('image-optimize', 'Failed to create image variant at: ')
+                    .$outputPath,
+                    __METHOD__
+                );
             }
 
             FileHelper::removeFile($outputPath);
+
         } else {
             Craft::error(
                 Craft::t('image-optimize', 'Failed to create image variant at: ')
@@ -455,7 +467,7 @@ class Optimize extends Component
                         'command'   => $thisImageProcessor['commandPath']
                             .' '
                             .$thisImageProcessor['commandOptions'],
-                        'installed' => file_exists($thisImageProcessor['commandPath']),
+                        'installed' => is_file($thisImageProcessor['commandPath']),
                     ];
                 }
             }
@@ -487,7 +499,7 @@ class Optimize extends Component
                         'command'   => $thisVariantCreator['commandPath']
                             .' '
                             .$thisVariantCreator['commandOptions'],
-                        'installed' => file_exists($thisVariantCreator['commandPath']),
+                        'installed' => is_file($thisVariantCreator['commandPath']),
                     ];
                 }
             }