+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # ImageOptimize Changelog
 
+## 1.3.5 - 2017.12.06
+### Changed
+* Switched from `file_exists()` to `is_file()` for detecting image processor and variant commands
+* Fixed an issue with Optimized Image Variant generation on certain setups
+* The OptimizedImages field now displays protocol-relative images properly
+* Added better debug logging
+* Updated to require craftcms/cms `^3.0.0-RC1`
+
 ## 1.3.4 - 2017.12.03
 ### Changed
 * Fixed an issue where selecting image variants would result in the wrong aspect ratio